+++ apiserver/paasng/tests/paasng/platform/applications/test_feature_flag.py
@@ -39,7 +39,6 @@ class TestGetDefaultFeatureFlags:
                     "APPLICATION_DESCRIPTION": True,
                     "MODIFY_ENVIRONMENT_VARIABLE": True,
                     "ENABLE_BK_LOG_COLLECTOR": False,
-                    "USE_SHARED_BK_LOG_INDEX": False,
                     "TOGGLE_EGRESS_BINDING": False,
                     "ENABLE_PERSISTENT_STORAGE": False,
                 },
@@ -56,7 +55,6 @@ class TestGetDefaultFeatureFlags:
                     "APPLICATION_DESCRIPTION": True,
                     "MODIFY_ENVIRONMENT_VARIABLE": True,
                     "ENABLE_BK_LOG_COLLECTOR": False,
-                    "USE_SHARED_BK_LOG_INDEX": False,
                     "TOGGLE_EGRESS_BINDING": False,
                     "ENABLE_PERSISTENT_STORAGE": False,
                 },
@@ -73,7 +71,6 @@ class TestGetDefaultFeatureFlags:
                     "APPLICATION_DESCRIPTION": True,
                     "MODIFY_ENVIRONMENT_VARIABLE": True,
                     "ENABLE_BK_LOG_COLLECTOR": False,
-                    "USE_SHARED_BK_LOG_INDEX": False,
                     "TOGGLE_EGRESS_BINDING": False,
                     "ENABLE_PERSISTENT_STORAGE": False,
                 },
@@ -90,7 +87,6 @@ class TestGetDefaultFeatureFlags:
                     "APPLICATION_DESCRIPTION": True,
                     "MODIFY_ENVIRONMENT_VARIABLE": True,
                     "ENABLE_BK_LOG_COLLECTOR": False,
-                    "USE_SHARED_BK_LOG_INDEX": False,
                     "TOGGLE_EGRESS_BINDING": False,
                     "ENABLE_PERSISTENT_STORAGE": False,
                 },
@@ -114,7 +110,6 @@ class TestGetDefaultFeatureFlags:
                     "APPLICATION_DESCRIPTION": True,
                     "MODIFY_ENVIRONMENT_VARIABLE": True,
                     "ENABLE_BK_LOG_COLLECTOR": False,
-                    "USE_SHARED_BK_LOG_INDEX": False,
                     "TOGGLE_EGRESS_BINDING": False,
                     "ENABLE_PERSISTENT_STORAGE": True,
                 },
